+++ crates/starlight/src/gc/collector.rs
