+++ crates/starlight/src/gc/large_object_space.rs
