+++ crates/starlight/src/gc/space_bitmap.rs
